refactor(driver): extract EntryVerificationHold to driver/hold.rs (refactor 2.1b)
Second commit of the Phase 2.1 mechanical split. Moves the
EntryVerificationHold state machine (Phase 1.6's artifact) out of
driver/mod.rs into a dedicated sibling module.

- `crates/based-rollup/src/driver/hold.rs` (new, 199 lines):
  contains `EntryVerificationHold`, `DeferralResult`,
  `MAX_ENTRY_VERIFY_DEFERRALS`, and the full module comment.
  Self-contained: no `use` statements needed — the type is pure Rust
  (u64 / u32 / enum) with no alloy/reth dependencies.

- `crates/based-rollup/src/driver/mod.rs` (-195 lines → 5317 total):
  replaced the 194-line block with:
    ```rust
    mod hold;
    pub use hold::{DeferralResult, EntryVerificationHold, MAX_ENTRY_VERIFY_DEFERRALS};
    ```
  The `pub use` re-export preserves the `use super::*` pattern in
  `driver_tests.rs` — every test that references `EntryVerificationHold`,
  `DeferralResult`, or `MAX_ENTRY_VERIFY_DEFERRALS` continues to resolve.

- `cargo check`, `cargo clippy --all-features -- -D warnings`, and
  `cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
  change — `git blame` on any method still points to its original
  commit via the file rename chain (mod.rs → hold.rs).

Next: 2.1c extracts `PendingL1SubmissionQueue` + `PendingL1Group` +
`BlockEntryMix` + `TriggerMetadata` to `driver/pending_queue.rs`.
